Rename AnalyticsService file.
