+++ lib/handlers/check-order-status/handler.ts
@@ -6,7 +6,7 @@ import { DutchOrderEntity } from '../../entities'
 import { CheckOrderStatusHandlerMetricNames, powertoolsMetric } from '../../Metrics'
 import { BaseOrdersRepository } from '../../repositories/base'
 import { LimitOrdersRepository } from '../../repositories/limit-orders-repository'
-import { AnalyticsService } from '../../services/analytics-service'
+import { AnalyticsService } from '../../services/AnalyticsService'
 import { SfnInjector, SfnLambdaHandler, SfnStateInputOutput } from '../base'
 import { kickoffOrderTrackingSfn } from '../shared/sfn'
 import { ContainerInjected, RequestInjected } from './injector'
