+++ lib/handlers/check-order-status/service.ts
@@ -5,7 +5,7 @@ import { log } from '../../Logging'
 import { CheckOrderStatusHandlerMetricNames, wrapWithTimerMetric } from '../../Metrics'
 import { checkDefined } from '../../preconditions/preconditions'
 import { BaseOrdersRepository } from '../../repositories/base'
-import { AnalyticsServiceInterface } from '../../services/analytics-service'
+import { AnalyticsServiceInterface } from '../../services/AnalyticsService'
 import { ChainId } from '../../util/chain'
 import { metrics } from '../../util/metrics'
 import { SfnStateInputOutput } from '../base'
