+++ lib/handlers/post-order/index.ts
@@ -4,7 +4,7 @@ import { ethers } from 'ethers'
 import { CONFIG } from '../../Config'
 import { log } from '../../Logging'
 import { DutchOrdersRepository } from '../../repositories/dutch-orders-repository'
-import { AnalyticsService } from '../../services/analytics-service'
+import { AnalyticsService } from '../../services/AnalyticsService'
 import { OrderDispatcher } from '../../services/OrderDispatcher'
 import { UniswapXOrderService } from '../../services/UniswapXOrderService'
 import { SUPPORTED_CHAINS } from '../../util/chain'
