+++ test/unit/services/AnalyticsService.test.ts
@@ -3,7 +3,7 @@ import { OrderType } from '@uniswap/uniswapx-sdk'
 import { mock } from 'jest-mock-extended'
 import { DutchOrderEntity, ORDER_STATUS } from '../../../lib/entities'
 import { LimitOrder } from '../../../lib/models/LimitOrder'
-import { AnalyticsService } from '../../../lib/services/analytics-service'
+import { AnalyticsService } from '../../../lib/services/AnalyticsService'
 import { ChainId } from '../../../lib/util/chain'
 import { SDKDutchOrderFactory } from '../../factories/SDKDutchOrderV1Factory'
 import { QUOTE_ID, SIGNATURE, Tokens } from '../fixtures'
